Moved the get comments endpoint to comments route, /api/posts/<post_id>/comments to /comments/<post_id>
