+++ app/routes/comments.py
@@ -2,7 +2,8 @@ from flask import Blueprint, request, jsonify
 from flask import g
 from app.middleware.auth import require_auth
 from app.models.user import get_or_create_user
-from app.models.comment import create_comment
+from app.models.post import get_post_by_id
+from app.models.comment import create_comment, get_comments_for_post
 
 bp = Blueprint('comments', __name__)
 
@@ -31,3 +32,19 @@ def add_comment():
     except Exception as e:
         print(f"[ERROR]: Cannot create comment: {e}")
         return jsonify( {"error": "Failed to create comment"} ), 500
+    
+@bp.route('/comments/<int:post_id>', methods=['GET'])
+def get_comments(post_id):
+    post = get_post_by_id(post_id)
+    if not post:
+        return jsonify({"error": "Post not found"}), 404
+
+    comments_row = get_comments_for_post(post_id)
+    comments = []
+    for c in comments_row:
+        comments.append({
+            "text": c["comment_text"],
+            "author": c["username"]
+        })
+
+    return jsonify(comments), 200
